testing: add common.py with CUDA+XPU tf32 helpers, refactor common_cuda.py
Introduce torch/testing/_internal/common.py as a device-agnostic home for
TF32-related test utilities that work across both CUDA and XPU backends.

Changes:
- New file common.py:
  * tf32_off(): disables TF32 for cuBLAS/cuDNN (CUDA) and oneDNN/mkldnn (XPU)
  * tf32_on(): enables TF32 for both backends, lowers test precision threshold
  * tf32_enabled(): context manager to temporarily enable TF32 for both backends
  * tf32_on_and_off(): decorator that runs a test twice (TF32 on and off);
    trigger condition extended to include XPU via torch.xpu.is_tf32_supported()
    and device type 'xpu' in addition to existing CUDA logic
  * with_tf32_off(): decorator that forces TF32 off for both backends

- common_cuda.py: replace the five function bodies with re-exports from
  common.py so that all existing
  'from torch.testing._internal.common_cuda import tf32_*'
  call sites continue to work without modification.

Related: intel/torch-xpu-ops#2400
